refactor(in-app): create directories for each dialog
