+++ personalization-sdk/src/main/kotlin/com/personalization/features/inAppNotification/impl/InAppNotificationManagerImpl.kt
@@ -14,13 +14,14 @@ import androidx.fragment.app.FragmentManager
 import com.personalization.R
 import com.personalization.api.managers.InAppNotificationManager
 import com.personalization.errors.EmptyFieldError
-import com.personalization.inAppNotification.view.component.dialog.ALERT_DIALOG_TAG
-import com.personalization.inAppNotification.view.component.dialog.AlertDialog
-import com.personalization.inAppNotification.view.component.dialog.BOTTOM_SHEET_TAG
-import com.personalization.inAppNotification.view.component.dialog.BottomSheetDialog
-import com.personalization.inAppNotification.view.component.dialog.FullScreenDialog
-import com.personalization.inAppNotification.view.component.dialog.TOP_SHEET_DIALOG
-import com.personalization.inAppNotification.view.component.dialog.TopSheetDialog
+import com.personalization.inAppNotification.view.component.dialog.fullScreen.FULL_SCREEN_DIALOG_TAG
+import com.personalization.inAppNotification.view.component.dialog.alert.ALERT_DIALOG_TAG
+import com.personalization.inAppNotification.view.component.dialog.alert.AlertDialog
+import com.personalization.inAppNotification.view.component.dialog.bottomSheet.BOTTOM_SHEET_TAG
+import com.personalization.inAppNotification.view.component.dialog.bottomSheet.BottomSheetDialog
+import com.personalization.inAppNotification.view.component.dialog.fullScreen.FullScreenDialog
+import com.personalization.inAppNotification.view.component.dialog.topSheet.TOP_SHEET_DIALOG
+import com.personalization.inAppNotification.view.component.dialog.topSheet.TopSheetDialog
 import com.personalization.inAppNotification.view.component.snackbar.Snackbar
 import com.personalization.sdk.data.models.dto.popUp.DialogDataDto
 import com.personalization.sdk.data.models.dto.popUp.PopupDto
