+++ personalization-sdk/src/main/kotlin/com/personalization/inAppNotification/view/component/dialog/alert/AlertDialog.kt
@@ -1,4 +1,4 @@
-package com.personalization.inAppNotification.view.component.dialog
+package com.personalization.inAppNotification.view.component.dialog.alert
 
 import android.graphics.Color
 import android.graphics.drawable.ColorDrawable
@@ -9,6 +9,7 @@ import android.view.View
 import android.view.ViewGroup
 import com.personalization.databinding.AlertDialogBinding
 import com.personalization.inAppNotification.view.component.container.InAppViewContainer
+import com.personalization.inAppNotification.view.component.dialog.BaseInAppDialog
 import com.personalization.inAppNotification.view.component.utils.InAppConsts.BUTTON_NEGATIVE_COLOR_KEY
 import com.personalization.inAppNotification.view.component.utils.InAppConsts.BUTTON_NEGATIVE_TEXT_KEY
 import com.personalization.inAppNotification.view.component.utils.InAppConsts.BUTTON_POSITIVE_COLOR_KEY
