+++ personalization-sdk/src/main/kotlin/com/personalization/inAppNotification/view/component/dialog/fullScreen/FullScreenDialog.kt
@@ -1,16 +1,12 @@
-@file:Suppress("NewApi")
+package com.personalization.inAppNotification.view.component.dialog.fullScreen
 
-package com.personalization.inAppNotification.view.component.dialog
-
-import android.graphics.Color
-import android.graphics.drawable.ColorDrawable
 import android.os.Bundle
-import android.view.Gravity
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.personalization.databinding.TopSheetDialogBinding
+import com.personalization.databinding.FullScreenDialogBinding
 import com.personalization.inAppNotification.view.component.container.InAppViewContainer
+import com.personalization.inAppNotification.view.component.dialog.BaseInAppDialog
 import com.personalization.inAppNotification.view.component.utils.InAppConsts.BUTTON_NEGATIVE_COLOR_KEY
 import com.personalization.inAppNotification.view.component.utils.InAppConsts.BUTTON_NEGATIVE_TEXT_KEY
 import com.personalization.inAppNotification.view.component.utils.InAppConsts.BUTTON_POSITIVE_COLOR_KEY
@@ -19,14 +15,14 @@ import com.personalization.inAppNotification.view.component.utils.InAppConsts.IM
 import com.personalization.inAppNotification.view.component.utils.InAppConsts.MESSAGE_KEY
 import com.personalization.inAppNotification.view.component.utils.InAppConsts.TITLE_KEY
 
-const val TOP_SHEET_DIALOG = "TopSheetDialog"
+const val FULL_SCREEN_DIALOG_TAG = "FullScreenDialog"
 
-class TopSheetDialog : BaseInAppDialog() {
+class FullScreenDialog : BaseInAppDialog() {
 
-    private var _binding: TopSheetDialogBinding? = null
+    private var _binding: FullScreenDialogBinding? = null
     private val binding get() = _binding!!
 
-    class TopSheetViewContainer(binding: TopSheetDialogBinding) : InAppViewContainer {
+    class FullScreenDialogViewContainer(binding: FullScreenDialogBinding) : InAppViewContainer {
         override val backgroundImageView = binding.backgroundImageView
         override val imageContainer = binding.imageContainer
         override val titleTextView = binding.textContainer.title
@@ -38,32 +34,21 @@ class TopSheetDialog : BaseInAppDialog() {
         override val closeButton = binding.closeButton
     }
 
-
     override val container: InAppViewContainer by lazy {
-        TopSheetViewContainer(binding)
+        FullScreenDialogViewContainer(binding)
     }
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = TopSheetDialogBinding.inflate(inflater, container, false)
+        _binding = FullScreenDialogBinding.inflate(inflater, container, false)
         return binding.root
     }
 
     override fun onDestroyView() {
         super.onDestroyView()
         _binding = null
-    }
-
-    override fun onStart() {
-        super.onStart()
-        dialog?.window?.apply {
-            setLayout(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.WRAP_CONTENT)
-            setGravity(Gravity.TOP)
-            setBackgroundDrawable(ColorDrawable(Color.TRANSPARENT))
-            setDimAmount(0.5f)
-        }
     }
 
     companion object {
@@ -74,9 +59,9 @@ class TopSheetDialog : BaseInAppDialog() {
             buttonPositiveText: String?,
             buttonNegativeText: String?,
             buttonPositiveColor: Int?,
-            buttonNegativeColor: Int?,
-        ): TopSheetDialog {
-            val dialog = TopSheetDialog()
+            buttonNegativeColor: Int?
+        ): FullScreenDialog {
+            val dialog = FullScreenDialog()
             val args = Bundle().apply {
                 putString(TITLE_KEY, title)
                 putString(MESSAGE_KEY, message)
